+++ com.unity.netcode.gameobjects/Runtime/Connection/NetworkClient.cs
@@ -1,14 +1,7 @@
-using System.Collections.Generic;
 using UnityEngine;
 
 namespace Unity.Netcode
 {
-
-    public enum NetworkTopologyTypes
-    {
-        ClientServer,
-        DistributedAuthority
-    }
 
     /// <summary>
     /// A NetworkClient
@@ -61,9 +54,9 @@ namespace Unity.Netcode
         public NetworkObject PlayerObject;
 
         /// <summary>
-        /// The list of NetworkObject's owned by this client instance
+        /// The NetworkObject's owned by this client instance
         /// </summary>
-        public List<NetworkObject> OwnedObjects => IsConnected ? SpawnManager.GetClientOwnedObjects(ClientId) : new List<NetworkObject>();
+        public NetworkObject[] OwnedObjects => IsConnected ? SpawnManager.GetClientOwnedObjects(ClientId) : new NetworkObject[] { };
 
         internal NetworkSpawnManager SpawnManager { get; private set; }